refactor(gateway): extract config/channel-override helpers from run.py (slice 18 of #54962)
